arc-gcp/laboratories/laboratory-1-microservices/api/requirements.txt
falcon==2.0.0
google-cloud-firestore==1.6.0